requirements.txt
pydra[doc,dev]>=1.0a9  # core engine + pydra.tasks.common; [doc] pulls in matplotlib for show_workflow
pydra-tasks-mrtrix3      # MrGrid, MrConvert examples
fileformats-medimage     # Nifti1
nibabel                  # NIfTI file I/O inside VoxelStats task
nest-asyncio             # required to run async tasks inside a notebook
jupyter
pydra2app-xnat >=0.5
xnat4tests
jupyter
bash_kernel
